once hart and johhny complete their parts the app will now magically work
