+++ README.md
@@ -1,1 +1,5 @@
-
+# HackerNews Clone
+### Created by...
+* Hart
+* Johnny
+* Rob
