+++ src/components/App.js
@@ -1,10 +1,13 @@
-import React, { useState, useContext } from 'react'
-import FrontPageContext from '../store/front-page-context'
+import React, { useState } from 'react'
+
 import Header from './Header/Header'
+import InitialList from './InitialList/InitialList'
+import ListBySearchTerm from './ListBySearchTerm/ListBySearchTerm'
 import classes from './App.module.css'
 
 const App = () => {
 // ! state and helper func for Johnnys header component
+// ! once you set your header inputs onChange to this state the app will begin to work
  const [searchTerm, setSearchTerm] = useState('')
 
  const handleOnChange = (e) => {
@@ -12,33 +15,8 @@ const App = () => {
  }
 
 
-// ! state and functions for me and Hart below
-// all inital front page search results array from the fetch request are saved in this variable
-const FetchFrontPageContext = useContext(FrontPageContext)
+// ! Me and hart will handle everything below here
 
-
-
-const data = FetchFrontPageContext.map((item, index )=> {
-    return (
-        <>
-            <li className={classes.title} key={index}>
-                {item.title}
-            </li>
-            <a className={classes.link} key={item.objectID} href={item.url}>
-                {item.url}
-            </a>
-            <div className={classes.info} key={Math.random()}>
-                <span className={classes['info-items']}>Created at: {item.created_at}</span>
-                <span className={classes['info-items']}> | </span>
-                <span className={classes['info-items']}>Created by: {item.author}</span>
-                <span className={classes['info-items']}> | </span>
-                <span className={classes['info-items']}>Points: {item.points}</span>
-            </div>
-            <hr />
-
-        </>
-    )
-})
 
 
 
@@ -46,9 +24,9 @@ const data = FetchFrontPageContext.map((item, index )=> {
   return (
     <main className={classes.wrapper}>
         <Header searchTerm={searchTerm} onChange={ (e) => handleOnChange(e) } />
-        <ul>
-          { data.length === 0 ? <h1>Loading Results...</h1> : data }
-        </ul>
+        {searchTerm ? <ListBySearchTerm /> : <InitialList />}
+       
+        
     </main>
   )
 }
