+++ src/components/App.module.css
@@ -1,23 +1,24 @@
 :root {
-
+    --side-border: 5vw solid black;
+    --btm-border: 5vh solid black;
 }
 
-.info {
-    display: flex;
-    flex-direction: row;
-}
-.info-items {
-    margin: .5px 5px;
-}
-.link {
-    text-decoration: none;
-}
-.title {
-    list-style: none;
-}
 .wrapper {
-    border-left: 5vw solid black;
-    border-right: 5vw solid black;
-    border-bottom: 5vh solid black;
+    border-left: var(--side-border);
+    border-right: var(--side-border);
+    border-bottom: var(--btm-border);
     background-color: rgb(248, 242, 242);
+    animation: list-appear 1s ease-out forwards;
 }
+
+@keyframes list-appear {
+    from {
+      opacity: 0;
+      transform: translateY(3rem);
+    }
+  
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
