+++ src/components/Header/Header.js
@@ -1,7 +1,9 @@
 import React from 'react'
 import './Header.css'
 
-const Header = () => {
+const Header = (props) => {
+    // Johnny You dont need to touch the app component, I already passed your props to your component
+    console.log(props)
   return (
     <div>
        SPACE FOR JOHHNYS HEADER (:
