+++ src/components/ListBySearchTerm/ListBySearchTerm.js
@@ -1,0 +1,41 @@
+import React, { useEffect, useState } from 'react'
+import classes from './ListBySearchTerm.module.css'
+
+const ListBySearchTerm = ({ term }) => {
+    const [searchResults, setSearchResults] = useState([])
+    // ! Hart will have to set up the search term to work
+
+useEffect(() => {
+    fetch(`https://hn.algolia.com/api/v1/search?query=building&tags=story`)
+        .then(response => response.json())
+        .then(data => setSearchResults(data.hits))
+}, [])
+
+const data = searchResults.map((item, index )=> {
+    return (
+        <>
+            <li className={classes.title} key={index}>
+                {item.title}
+            </li>
+            <a className={classes.link} key={item.objectID} href={item.url}>
+                {item.url}
+            </a>
+            <div className={classes.info} key={Math.random()}>
+                <span className={classes['info-items']}>Points: <b>{item.points}</b></span>
+                <span className={classes['info-items']}> | </span>
+                <span className={classes['info-items']}>Created by: <b>{item.author}</b></span>
+                <span className={classes['info-items']}> | </span>
+                <span className={classes['info-items']}>Created at: <b>{item.created_at}</b></span>
+            </div>
+            <hr />
+        </>
+    )
+})
+
+
+  return (
+    <div>{ data.length === 0 ? <h1>Loading...</h1> : data }</div>
+  )
+}
+
+export default ListBySearchTerm
