+++ src/components/ListBySearchTerm/ListBySearchTerm.module.css
@@ -1,0 +1,22 @@
+:root {
+    --title-font-weight: 600;
+    --info-items-pad-small: 3px 5px;
+}
+
+.info {
+    display: flex;
+    flex-direction: row;
+}
+.info-items {
+    margin: var(--info-items-pad-small);
+}
+.link {
+    text-decoration: none;
+    padding-bottom: 1rem;
+}
+.title {
+    list-style: none;
+    padding-bottom: .5rem;
+    font-weight: var(--title-font-weight);
+    font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
+}
